Avatar manager (WIP)
