+++ server/users.ts
@@ -179,6 +179,10 @@ function isUsernameKnown(name: string) {
 	return false;
 }
 
+function isUsername(name: string) {
+	return /[A-Za-z0-9]/.test(name.charAt(0)) && /[A-Za-z]/.test(name) && !name.includes(',');
+}
+
 function isTrusted(userid: ID) {
 	if (globalAuth.has(userid)) return userid;
 	for (const room of Rooms.global.chatRooms) {
@@ -1061,9 +1065,7 @@ export class User extends Chat.MessageContext {
 		this.registered = true;
 		if (!isMerge) this.tempGroup = globalAuth.get(this.id);
 
-		if (Config.customavatars && Config.customavatars[this.id]) {
-			this.avatar = Config.customavatars[this.id];
-		}
+		Users.Avatars?.handleLogin(this);
 
 		const groupInfo = Config.groups[this.tempGroup];
 		this.isStaff = !!(groupInfo && (groupInfo.lock || groupInfo.root));
@@ -1701,8 +1703,10 @@ export const Users = {
 	getExact: getExactUser,
 	findUsers,
 	Auth,
+	Avatars: null as typeof import('./chat-commands/avatars').Avatars | null,
 	globalAuth,
 	isUsernameKnown,
+	isUsername,
 	isTrusted,
 	SECTIONLEADER_SYMBOL,
 	PLAYER_SYMBOL,
